requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
networkx
pyvis
scikit-learn
numpy
torch